Add files via upload
A very basic GPA calculator I built while taking CIT-144. This one was a lot of fun to write! Please do not copy line for line if you are using this program for your own studies, you will learn much more if you use it as a blueprint and integrate your own code into it instead of pasting it line by line just to pass.